# contracts/v1_api_swagger.json
{
  "openapi": "3.0.1",
...
  "info": {
    "title": "CipherSwarm Agent API",
    "description": "The CipherSwarm Agent API is used to allow agents to connect to the CipherSwarm server.",
    "version": "1.3",
    "license": {
      "name": "Mozilla Public License Version 2.0",
      "url": "https://www.mozilla.org/en-US/MPL/2.0/"
    }
  },
  "servers": [
    {
      "url": "https://{defaultHost}",
      "description": "The production server",
      "variables": {
        "defaultHost": {
          "default": "www.example.com"
        }
      }
    },
    {
      "url": "http://{hostAddress}:{hostPort}",
      "description": "The insecure server",
      "variables": {
        "hostAddress": {
          "default": "localhost"
        },
        "hostPort": {
          "default": "8080"
        }
      }
    }
  ],
  "security": [
    {
      "bearer_auth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ErrorObject": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/client/authenticate": {
      "get": {
        "summary": "Authenticate Client",
        "tags": [
          "Client"
        ],
        "description": "Authenticates the client. This is used to verify that the client is able to connect to the server.",
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "operationId": "authenticate",
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "authenticated": true,
                      "agent_id": 2624
                    }
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "authenticated": {
                      "type": "boolean"
                    },
                    "agent_id": {
                      "type": "integer",
                      "format": "int64"
                    }
                  },
                  "required": [
                    "authenticated",
                    "agent_id"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Bad credentials"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          }
        }
      }
    }
  }
}